requirements.txt
aiosmtpd==1.4.6
alembic==1.17.0
anyio==4.11.0
atpublic==6.0.2
attrs==25.4.0
babel==2.17.0
black==25.9.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
croniter==6.0.0
dnspython==2.8.0
elastic-transport==9.2.0
elasticsearch==9.2.0
email-validator==2.3.0
Flask==3.1.2
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-translate==3.23.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
langdetect==1.0.9
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.0
pytokens==0.3.0
pytz==2025.2
redis==7.0.1
requests==2.32.5
rq==2.6.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
